Rename test db folder to memory_store
