+++ p2panda-rs/src/test_utils/memory_store/stores/entry.rs
@@ -12,7 +12,7 @@ use crate::operation::EncodedOperation;
 use crate::schema::SchemaId;
 use crate::storage_provider::error::EntryStorageError;
 use crate::storage_provider::traits::{AsStorageLog, EntryStore};
-use crate::test_utils::db::{MemoryStore, StorageEntry, StorageLog};
+use crate::test_utils::memory_store::{MemoryStore, StorageEntry, StorageLog};
 
 /// Implement `EntryStore` trait on `MemoryStore`
 #[async_trait]
